requirements.txt
# Core Dependencies
streamlit==1.50.0
streamlit-option-menu==0.4.0

# Data Processing & Visualization
pandas==2.3.3
numpy==2.2.6
plotly==6.5.0
altair==5.5.0
pyarrow==21.0.0

# Image Processing
pillow==11.3.0
scipy==1.15.3
PyWavelets==1.8.0
opencv-python
scikit-image

# Database
psycopg2-binary==2.9.11
python-dotenv==1.1.1

# Security & Cryptography
cryptography==46.0.2
pycparser==2.23
reedsolo

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
charset-normalizer==3.4.3
certifi==2025.10.5

# Infrastructure
click==8.3.0
colorama==0.4.6
blinker==1.9.0
tornado==6.5.2
protobuf==6.32.1
cachetools==6.2.0
attrs==25.4.0

# Testing
pytest
torch
torchvision
